Implement Relationship- and PolyItemType mapping
Implemented and tested the mapping of relationship- and poly-ItemTypes
to DotArrow and/or DotClass. If some of the corresponding Items (either
Morphae for poly or source/related ItemTypes for relations) did not exist
yet as DotClasses (because they are either relationships themselves or
did not begin with the provided prefix), those will be created as well.
Also, if a relationship is a null-relationship, a corresponding class
will be created, together with different arrowheads (UML aggregation).