Implement MapRelsAsClasses functionality
Implemented functionality that deals with the arguments option "-r",
meaning the user wants to have Aras relationship ItemTypes mapped as
classes in the class diagram as well.
This functionality creates a DotClass from the relationship ItemType and
links it to source_id and related_id via arrows originating from it.