Add support for multiple prefixes and wider search
It is now possible to enter multiple prefixes that shall be taken into
consideration when looking for ItemTypes.
Additionally, I introduced another command line argument. It allows the
user to specify whether the prefixes and package name ItemTypes shall be
combined like a union (all ItemTypes that start with one of the prefixes
OR are in the given package) or an intesection (all ItemTypes that
start with one of the prefixes AND are in the given package).