Remove redundant class definitions
Due to a bug, multiple classes of the same name were created if that
name included spaces. The reason for this was that we need to 'nest' the
class name in quotation marks if it contains spaces (otherwise dot will
give syntax errors). But when we were searching for already existing
classes in the code, we only ever searched for the class name without
the quotations marks around it. This led the program to think it found a
new class that it needed to create, when, in fact, the desired class was
already there, simply with a slightly modified name.
This behaviour is fixed now and the program now tries to find classes
both using the raw name and the name in quotation marks.